extract Markdown config types to separate file + add config.markdown.hooks.onBrokenLink
